Stops training from crashing
Fixes shape problems in net when training.

The problem was actually something with Tensorflow itself, where the Lambda layers were theorically formed correctly, but the first Action in the action layer was receiving the output from the last built Lambda layer, which was naturally outputting a different shape.

I don't know why this happened, but encapsulating the instancing of Lambda Layers with _builds_connections_layer() seems to fix it.