ASNet Training
Applies the techniques described in the paper to planning. That is, first explores states reached by the execution of the ASNet, and then expand on those states with the Teacher policy to make sure we will have optimal states in training.

Also sets arbitrary batch sizes and training epoch values, as these are not specified in the paper.

The training of an ASNet capable of solving the simplest Cubeworld problem was a success!